+++ src/parser/inline/autolink.ts
@@ -5,6 +5,7 @@ import { email } from './autolink/email';
 import { channel } from './autolink/channel';
 import { account } from './autolink/account';
 import { hashtag } from './autolink/hashtag';
+import { hashref } from './autolink/hashref';
 
 export const autolink: AutolinkParser = union([
   uri,
@@ -12,4 +13,5 @@ export const autolink: AutolinkParser = union([
   channel,
   account,
   hashtag,
+  hashref,
 ]);