+++ src/parser/inline/autolink/hashtag.test.ts
@@ -9,6 +9,7 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('#')), [['#'], '']);
       assert.deepStrictEqual(inspect(parser('# ')), [['#'], ' ']);
+      assert.deepStrictEqual(inspect(parser('#1')), undefined);
       assert.deepStrictEqual(inspect(parser('#a#')), undefined);
       assert.deepStrictEqual(inspect(parser('#\\')), [['#'], '\\']);
       assert.deepStrictEqual(inspect(parser('#\\ ')), [['#'], '\\ ']);
@@ -21,14 +22,15 @@ describe('Unit: parser/inline/autolink/hashtag', () => {
       assert.deepStrictEqual(inspect(parser('#{#}')), [['#'], '{#}']);
       assert.deepStrictEqual(inspect(parser('#{a}')), [['#'], '{a}']);
       assert.deepStrictEqual(inspect(parser('#　')), [['#'], '　']);
+      assert.deepStrictEqual(inspect(parser('a#1')), undefined);
       assert.deepStrictEqual(inspect(parser('a#b')), [['a#'], 'b']);
+      assert.deepStrictEqual(inspect(parser('a##1')), undefined);
       assert.deepStrictEqual(inspect(parser('a##b')), [['a##'], 'b']);
       assert.deepStrictEqual(inspect(parser('あ#b')), [['あ#'], 'b']);
       assert.deepStrictEqual(inspect(parser(' #a')), undefined);
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('#0')), [['<a class="hashtag" rel="noopener">#0</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#a')), [['<a class="hashtag" rel="noopener">#a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('#a ')), [['<a class="hashtag" rel="noopener">#a</a>'], ' ']);
       assert.deepStrictEqual(inspect(parser('#a\n')), [['<a class="hashtag" rel="noopener">#a</a>'], '\n']);